Add files via upload
This project analyzed the happiness scores among 149 countries using different variables. The main objective is to see if COVID-19 had a significant impact on people's overall happiness around the world. The study is inspired by the World Happiness Report 2021. This project was submitted to the University of Washington as a final requirement.